+++ core/src/main/scala/org/apache/spark/security/HTTPHelper.scala
@@ -20,7 +20,7 @@ package org.apache.spark.security
 import java.io.{BufferedReader, File, InputStreamReader}
 import java.security.cert.X509Certificate
 
-import scala.annotation.tailrec
+import scala.util.Try
 import scala.util.parsing.json.JSON
 
 import org.apache.http.client.HttpClient
@@ -30,6 +30,10 @@ import org.apache.http.impl.client.HttpClients
 import org.apache.http.ssl.{SSLContextBuilder, TrustStrategy}
 
 import org.apache.spark.internal.Logging
+
+
+
+
 
 object HTTPHelper extends Logging{
 
@@ -54,14 +58,14 @@ object HTTPHelper extends Logging{
   def executePost(requestUrl: String,
                   parentField: String,
                   headers: Option[Seq[(String, String)]],
-                  entity: Option[String] = None): Map[String, Any] = {
+                  entity: Option[String] = None): Try[Map[String, Any]] = {
     val post = new HttpPost(requestUrl)
 
     getContentFromResponse(post, parentField, headers, entity)
   }
   def executeGet(requestUrl: String,
                  parentField: String,
-                 headers: Option[Seq[(String, String)]]): Map[String, Any] = {
+                 headers: Option[Seq[(String, String)]]): Try[Map[String, Any]] = {
     val get = new HttpGet(requestUrl)
     getContentFromResponse(get, parentField, headers)
   }
@@ -69,34 +73,38 @@ object HTTPHelper extends Logging{
   private def getContentFromResponse(uriRequest: HttpUriRequest,
                                      parentField: String,
                                      headers: Option[Seq[(String, String)]],
-                                     entities: Option[String] = None): Map[String, Any] = {
+                                     entities: Option[String] = None): Try[Map[String, Any]] =
+    Try {
+      headers.foreach(
+        head => head.foreach { case (head, value) => uriRequest.addHeader(head, value) }
+      )
 
-    headers.map(head => head.foreach { case (head, value) => uriRequest.addHeader(head, value) })
+      entities.foreach(
+        entity => uriRequest.asInstanceOf[HttpPost].setEntity(new StringEntity(entity))
+      )
 
-    entities.map(entity => uriRequest.asInstanceOf[HttpPost].setEntity(new StringEntity(entity)))
+      val client = secureClient match {
+        case Some(secureClient) =>
+          logDebug(s"Using secure client")
+          secureClient
+        case _ => logDebug(s"Using non secure client")
+          clientNaive
+      }
 
-    val client = secureClient match {
-      case Some(secureClient) =>
-        logDebug(s"Using secure client")
-        secureClient
-      case _ => logDebug(s"Using non secure client")
-        clientNaive
-    }
+      val response = client.execute(uriRequest)
 
-    val response = client.execute(uriRequest)
+      val rd = new BufferedReader(
+        new InputStreamReader(response.getEntity().getContent()))
 
-    val rd = new BufferedReader(
-      new InputStreamReader(response.getEntity().getContent()))
-
-    val json = JSON.parseFull(Stream.continually(rd.readLine()).takeWhile(_ != null).mkString).
-      get.asInstanceOf[Map[String, Any]]
-    logTrace(s"getFrom Vault ${json.mkString("\n")}")
-    if(response.getStatusLine.getStatusCode != 200) {
-      val errors = json("errors").asInstanceOf[List[String]].mkString("\n")
-      throw new RuntimeException(errors)
-    }
-    else {
-      json(parentField).asInstanceOf[Map[String, Any]]
-    }
+      val json = JSON.parseFull(Stream.continually(rd.readLine()).takeWhile(_ != null).mkString).
+        get.asInstanceOf[Map[String, Any]]
+      logTrace(s"getFrom Vault ${json.mkString("\n")}")
+      if (response.getStatusLine.getStatusCode != 200) {
+        val errors = json("errors").asInstanceOf[List[String]].mkString("\n")
+        throw new RuntimeException(errors)
+      }
+      else {
+        json(parentField).asInstanceOf[Map[String, Any]]
+      }
   }
 }