+++ testsAT/src/test/java/com/stratio/spark/tests/specs/BaseSpec.java
@@ -1,0 +1,11 @@
+
+package com.stratio.spark.tests.specs;
+
+public class BaseSpec {
+
+    protected Common commonspec;
+
+    public Common getCommonSpec() {
+        return this.commonspec;
+    }
+}